requirements.txt
langchain>=0.3
langgraph>=0.2
langchain-groq>=0.2
langchain-google-genai>=2.0
langchain-huggingface>=0.1
langchain-community>=0.3
faiss-cpu>=1.9
sentence-transformers>=3.0
streamlit>=1.40
python-dotenv>=1.0
